Refactor run script to use helper functions
This is a nice reduction in duplicate code while still keeping things
readable in the end.

Although this isn't a proper 100% perfect solution. We're filting out
these helper functions in the task's help menu but technically you can
still run them directly, such as: run _dc

But still, this is very much worth it. Just don't run the helper functions
directly and it's all good!