Add public/.keep file to be tracked by git
The idea here is that we always want to ensure that we have a public/
directory commit to version control.

But the .gitignore file will ensure all of the files in there are
ignored except for the .keep file.

This will ensure the public/ directory is always owned by the correct
user since Docker will never have to create it by default if it doesn't
exist.