Add initializers.py to define frequently used code
Redis was never an extension in the way we were using it. In larger
projects I often want a separate file to define things that I may use
all across a project that isn't a Flask extension.

Redis is a good example in this case, but another one is using the
"Faker" library where you define it once and use it everywhere.