Explicitly set DEBUG config option
By default Flask will set DEBUG based on the value of FLASK_DEBUG but
for something as important as debug mode being enabled or not we should
be more explicit and define DEBUG ourselves.

That demonstrates intent in what we want it to be set to.

We'll also reference DEBUG instead of FLASK_DEBUG for the nav bar
because FLASK_DEBUG is optional where as DEBUG will always be set.